stream 1.5 - release new tests for generic signature (add javac comparison)

+++ org.eclipse.jdt.core.tests.compiler/src/org/eclipse/jdt/core/tests/compiler/regression/AbstractRegressionTest.java
@@ -480,7 +480,9 @@ public abstract class AbstractRegressionTest extends AbstractCompilerTest implem
 		}
 		// clean up output dir
 		File outputDir = new File(OUTPUT_DIR);
-		Util.flushDirectoryContent(outputDir);
-		outputDir.delete();
+		if (outputDir.exists()) {
+			Util.flushDirectoryContent(outputDir);
+			outputDir.delete();
+		}
 	}
 }
